+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - The [openLinks](./README.md#openlinks) extension now supports more parser functions when used on a MediaWiki site
 
+**Changed**
+
+- The [openLinks](./README.md#openlinks) extension is now only triggered by a main button click on text with underline
+
 ## 2.23.1
 
 *2025-03-15*